+++ script/workspace/workspace.lua
@@ -391,17 +391,22 @@ function m.findUrisByFilePath(path)
         return resultCache[path].results, resultCache[path].posts
     end
     tracy.ZoneBeginN('findUrisByFilePath #1')
+    local strict = config.get 'Lua.runtime.pathStrict'
     local results = {}
     local posts = {}
     for uri in files.eachFile() do
         if not uri:find(lpath, 1, true) then
             goto CONTINUE
         end
+        local relat = m.getRelativePath(uri)
         local pathLen = #path
-        local curPath = furi.decode(uri)
+        local curPath = relat
         local curLen  = #curPath
         local seg = curPath:sub(curLen - pathLen, curLen - pathLen)
         if seg == '/' or seg == '\\' or seg == '' then
+            if strict and seg ~= '' then
+                goto CONTINUE
+            end
             local see = curPath:sub(curLen - pathLen + 1, curLen)
             if see == path then
                 results[#results+1] = uri