+++ changelog.md
@@ -1,5 +1,9 @@
 # changelog
 
+## 2.5.0
+* `NEW` setting `Lua.runtime.pathStrict`
+* `CHG` no longer asks to trust plugin in VSCode, because VSCode already provides the workspace trust feature
+
 ## 2.4.7
 `2021-10-27`
 * `FIX` [#762](https://github.com/sumneko/lua-language-server/issues/762)